Versión clase

+++ registro.php
@@ -1,7 +1,6 @@
 <?php
     include('db.php');
     
-    // File upload path
     $targetDir = "uploads/";
     $filename = $_FILES["file"]["name"];
     $targetFilePath = $targetDir . $filename;
@@ -11,19 +10,42 @@
     $password = $_POST['password'];
     $email = $_POST['email'];
 
-    $sql = 'INSERT INTO `usuarios`.`users` (`id` ,`nickname` ,`password`, `email`, `avatar`)VALUES (NULL , "'.$nickname.'", "'.$password.'", "'.$email.'", "'.$filename.'")';
-    
     if(isset($_POST["submit"])){
-        $tempname = $_FILES["file"]["tmp_name"];
-        $folder = "uploads/".$filename;
+        if(move_uploaded_file($_FILES["file"]["tmp_name"], $targetFilePath))
+        {
+            $imageData = base64_encode(file_get_contents($targetFilePath));
+            $src = 'data:'.$fileType.';base64,' . $imageData;
 
-        if(move_uploaded_file($_FILES["file"]["tmp_name"], $targetFilePath)) {
-            $consulta = mysqli_query($con,$sql);
+            $sql='SELECT * FROM `users` WHERE 1';
+            $consultaUsersExistentes = mysqli_query($con,$sql);
 
-            if (!$consulta) {
-                die('<br>No se ha podido realizar el insert');
+            $userExiste = false;
+            if (trim($nickname) != "" && trim($password) != "") {
+                while($fila=$consultaUsersExistentes->fetch_assoc()) {
+                    if ($nickname == $fila['nickname'] AND $password == $fila['password']) {
+                        $userExiste = true;
+                    }
+                }
             } else {
-                echo '<br>Se realizó el insert';
+                die('El nombre o contraseña están vacíos');
+            }
+
+            if (!$userExiste) {
+                $sql = 'INSERT INTO `usuarios`.`users` (`id` ,`nickname` ,`password`, `email`, `avatar`)VALUES (NULL , "'.$nickname.'", "'.$password.'", "'.$email.'", "'.$src.'")';
+                $consulta = mysqli_query($con,$sql);
+
+                if (!$consulta) {
+                    die('<br>No se ha podido realizar el insert');
+                } else {
+                    echo '<br>Se realizó el insert';
+                }
+            } else {
+                echo 'El usuario ya existe';
+                echo '<button >Volver atrás</button>'
+
+                echo "<form action='index.html'>";
+                    echo "<button type='submit' value='Login'>Volver atrás</button>"
+                echo "</form>";
             }
         } else {
             echo 'no se pudo mover a uploads';
@@ -38,6 +60,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
+    <link rel="stylesheet" href="styles.css">
     </head>
         <body>
             <form action="login.html">
